chore: fix failing events disabled test
